+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 1.9.3-beta.6
 
+- `TypeInfoEntityExtension`:
+  - `entityType`: also handle `List<Entity>`, returning the `List` generic type.
+
 - args: ^2.7.0
 - postgres: ^3.5.6
 - archive: ^4.0.7