+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 - `TypeInfoEntityExtension`, `TypeReflectionEntityExtension`:
   - `entityType`: also handle `List<E>`, returning the `List` generic type (`E`).
 
+- `TypeInfoEntityExtension`:
+  - Added `toCastedList`.
+
+- `DBSQLAdapter`:
+  - `_checkDBTableScheme`:
+    - Separate references and collection references in `referenceFields` and `collectionReferenceFields`. 
+  - `_DBTableCheck`: added field `missingCollectionReferenceColumns`.
+
+- `EntityHandler`:
+  - `resolveFieldsValues`: ensure that `List<E>` fields are casted to the list, using `entityType.toCastedList(val)`.
+
 - args: ^2.7.0
 - postgres: ^3.5.6
 - archive: ^4.0.7